image-tools: Init with buildSpecifiedLayers
This isn't as "neat" as the automatic layering, as it will produce "fat"
layers, but this allows absolute control over layers.

It will produce one output per layer, but this is both a drawback and a
benefit. At the cost of some storage, we have a "memoized" set of
previous layers the build can start back from.

This means that in the scenario with layers `[ bash ruby app ]`, changing
only the app layer will use the previously built image instead of
rebuilding the whole image again.

The only drawback, compared to `dockerTools.buildLayeredImage` is that
there is no discrete layers re-used across different unrelated builds.
The builds have to be somehow related to have sharing. Though note that
in `[ bash ruby ]`, as long as they are uncustomized, those layers would
be re-used across unrelated images.